move opensearch to consumption
