+++ framework/src/consumption/lib/opensearch/index.ts
@@ -1,0 +1,2 @@
+export * from './opensearch';
+export * from './opensearch-props';
