+++ framework/src/consumption/lib/opensearch/opensearch-props.ts
