+++ framework/src/consumption/lib/opensearch/opensearch.ts
