+++ framework/src/consumption/lib/opensearch/resources/lambda/helper.ts
