+++ framework/src/consumption/lib/opensearch/resources/lambda/opensearch-api.ts
