+++ framework/src/consumption/lib/opensearch/resources/lambda/package.json
