+++ framework/src/consumption/lib/redshift/index.ts
@@ -1,5 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './redshift-data';
-export * from './redshift-data-props';
+export * from './redshift-serverless';
+export * from './redshift';
