+++ framework/src/consumption/lib/redshift/redshift-serverless/index.ts
