+++ framework/src/consumption/lib/redshift/redshift-serverless/redshift-serverless-namespace-props.ts
