+++ framework/src/consumption/lib/redshift/redshift-serverless/redshift-serverless-workgroup-props.ts
