+++ framework/src/consumption/lib/redshift/redshift/index.ts
@@ -1,6 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './redshift';
-export * from './athena';
-export * from './opensearch';
+export * from './redshift-data';
+export * from './redshift-data-props';
