+++ framework/src/consumption/lib/redshift/redshift/redshift-data-props.ts
