+++ framework/src/consumption/lib/redshift/resources/RedshiftServerlessNamespace/index.mjs
