+++ framework/src/consumption/lib/redshift/resources/RedshiftServerlessNamespace/package.json
