+++ framework/src/storage/lib/index.ts
@@ -7,5 +7,3 @@ export * from './data-lake-storage-props';
 export * from './analytics-bucket';
 export * from './analytics-bucket-props';
 export * from './access-logs-bucket';
-export * from './opensearch/opensearch-props';
-export * from './opensearch/opensearch';
