--- framework/src/storage/lib/opensearch/index.ts
@@ -1,2 +1,0 @@
-export { OpensearchCluster } from './opensearch';
-export { OpensearchProps, OpensearchNodes } from './opensearch-props';
